requirements.txt
itchat>=1.3.10
requests>=2.25.1
certifi>=2021.5.30
Pillow>=10.0.0
qrcode>=6.1
watchdog>=2.1.0
